cant get state
